Extract the alias tracker
