+++ datamapper/query/__init__.py
@@ -1,0 +1,3 @@
+from .query import Query
+
+__all__ = ["Query"]
